requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
Cython==0.29.28
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-pubsub==2.17.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
Pillow==9.5.0
ply==3.11
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
